+++ lib/aws-record/record/table_config.rb
@@ -33,6 +33,17 @@ module Aws
     #     t.write_capacity_units 5
     #   end
     #
+    # @example A basic model with pay per request billing.
+    #   class Model
+    #     include Aws::Record
+    #     string_attr :uuid, hash_key: true
+    #   end
+    #
+    #   table_config = Aws::Record::TableConfig.define do |t|
+    #     t.model_class Model
+    #     t.billing_mode "PAY_PER_REQUEST"
+    #   end
+    #
     # @example Running a conditional migration on a basic model.
     #   table_config = Aws::Record::TableConfig.define do |t|
     #     t.model_class Model
@@ -59,7 +70,9 @@ module Aws
     #       :title,
     #       hash_key:  :forum_uuid,
     #       range_key: :post_title,
-    #       projection_type: "ALL"
+    #       projection: {
+    #         projection_type: "ALL"
+    #       }
     #     )
     #   end
     #
@@ -106,7 +119,12 @@ module Aws
         #   * +#write_capacity_units+ Sets the write capacity units for the
         #     index.
         # * +#ttl_attribute+ Sets the attribute ID to be used as the TTL
-        #     attribute, and if present, TTL will be enabled for the table.
+        #   attribute, and if present, TTL will be enabled for the table.
+        # * +#billing_mode+ Sets the billing mode, with the current supported
+        #   options being "PROVISIONED" and "PAY_PER_REQUEST". If using
+        #   "PAY_PER_REQUEST" you must not set provisioned throughput values,
+        #   and if using "PROVISIONED" you must set provisioned throughput
+        #   values. Default assumption is "PROVISIONED".
         #
         # @example Defining a migration with a GSI.
         #   class Forum
@@ -151,6 +169,7 @@ module Aws
       def initialize
         @client_options = {}
         @global_secondary_indexes = {}
+        @billing_mode = "PROVISIONED" # default
       end
 
       # @api private
@@ -195,6 +214,11 @@ module Aws
         end
       end
 
+      # @api private
+      def billing_mode(mode)
+        @billing_mode = mode
+      end
+
       # Performs a migration, if needed, against the remote table. If
       # +#compatible?+ would return true, the remote table already has the same
       # throughput, key schema, attribute definitions, and global secondary
@@ -209,13 +233,7 @@ module Aws
           else
             # Gotcha: You need separate migrations for indexes and throughput
             unless _throughput_equal(resp)
-              @client.update_table(
-                table_name: @model_class.table_name,
-                provisioned_throughput: {
-                  read_capacity_units: @read_capacity_units,
-                  write_capacity_units: @write_capacity_units
-                }
-              )
+              @client.update_table(_update_throughput_opts(resp))
               @client.wait_until(
                 :table_exists,
                 table_name: @model_class.table_name
@@ -327,12 +345,19 @@ module Aws
 
       def _create_table_opts
         opts = {
-          table_name: @model_class.table_name,
-          provisioned_throughput: {
+          table_name: @model_class.table_name
+        }
+        if @billing_mode == "PROVISIONED"
+          opts[:provisioned_throughput] = {
             read_capacity_units: @read_capacity_units,
             write_capacity_units: @write_capacity_units
           }
-        }
+        elsif @billing_mode == "PAY_PER_REQUEST"
+          opts[:billing_mode] = @billing_mode
+        else
+          raise ArgumentError, "Unsupported billing mode #{@billing_mode}"
+        end
+          
         opts[:key_schema] = _key_schema
         opts[:attribute_definitions] = _attribute_definitions
         gsi = _global_secondary_indexes
@@ -340,6 +365,54 @@ module Aws
           opts[:global_secondary_indexes] = gsi 
         end
         opts
+      end
+
+      def _add_global_secondary_index_throughput(opts, resp_gsis)
+        gsis = resp_gsis.map do |g|
+          g.index_name
+        end
+        gsi_updates = []
+        gsis.each do |index_name|
+          lgsi = @global_secondary_indexes[index_name.to_sym]
+          gsi_updates << {
+            update: {
+              index_name: index_name,
+              provisioned_throughput: lgsi.provisioned_throughput
+            }
+          }
+        end
+        opts[:global_secondary_index_updates] = gsi_updates
+        true
+      end
+
+      def _update_throughput_opts(resp)
+        if @billing_mode == "PROVISIONED"
+          opts = {
+            table_name: @model_class.table_name,
+            provisioned_throughput: {
+              read_capacity_units: @read_capacity_units,
+              write_capacity_units: @write_capacity_units
+            }
+          }
+          # special case: we have global secondary indexes existing, and they
+          # need provisioned capacity to be set within this call
+          if !resp.table.billing_mode_summary.nil? &&
+              resp.table.billing_mode_summary.billing_mode == "PAY_PER_REQUEST"
+            opts[:billing_mode] = @billing_mode
+            if resp.table.global_secondary_indexes
+              resp_gsis = resp.table.global_secondary_indexes
+              _add_global_secondary_index_throughput(opts, resp_gsis)
+            end
+          end # else don't include billing mode
+          opts
+        elsif @billing_mode == "PAY_PER_REQUEST"
+          {
+            table_name: @model_class.table_name,
+            billing_mode: "PAY_PER_REQUEST"
+          }
+        else
+          raise ArgumentError, "Unsupported billing mode #{@billing_mode}"
+        end
       end
 
       def _update_index_opts(resp)
@@ -369,22 +442,25 @@ module Aws
           gsi[:key_schema].each do |k|
             attributes_referenced.add(k[:attribute_name])
           end
-          # This may be a problem, check if I can maintain symbols.
-          lgsi = @global_secondary_indexes[index_name.to_sym]
-          gsi[:provisioned_throughput] = lgsi.provisioned_throughput
+          if @billing_mode == "PROVISIONED"
+            lgsi = @global_secondary_indexes[index_name.to_sym]
+            gsi[:provisioned_throughput] = lgsi.provisioned_throughput
+          end
           gsi_updates << {
             create: gsi
           }
         end
-        update_candidates.each do |index_name|
-          # This may be a problem, check if I can maintain symbols.
-          lgsi = @global_secondary_indexes[index_name.to_sym]
-          gsi_updates << {
-            update: {
-              index_name: index_name,
-              provisioned_throughput: lgsi.provisioned_throughput
+        # we don't currently update anything other than throughput
+        if @billing_mode == "PROVISIONED"
+          update_candidates.each do |index_name|
+            lgsi = @global_secondary_indexes[index_name.to_sym]
+            gsi_updates << {
+              update: {
+                index_name: index_name,
+                provisioned_throughput: lgsi.provisioned_throughput
+              }
             }
-          }
+          end
         end
         attribute_definitions = _attribute_definitions
         incremental_attributes = attributes_referenced.map do |attr_name|
@@ -438,13 +514,18 @@ module Aws
       end
 
       def _throughput_equal(resp)
-        expected = resp.table.provisioned_throughput.to_h
-        actual = {
-          read_capacity_units: @read_capacity_units,
-          write_capacity_units: @write_capacity_units
-        }
-        actual.all? do |k,v|
-          expected[k] == v
+        if @billing_mode == "PAY_PER_REQUEST"
+          !resp.table.billing_mode_summary.nil? &&
+            resp.table.billing_mode_summary.billing_mode == "PAY_PER_REQUEST"
+        else
+          expected = resp.table.provisioned_throughput.to_h
+          actual = {
+            read_capacity_units: @read_capacity_units,
+            write_capacity_units: @write_capacity_units
+          }
+          actual.all? do |k,v|
+            expected[k] == v
+          end
         end
       end
 
@@ -498,10 +579,17 @@ module Aws
           remote_key_schema = rgsi.key_schema.map { |i| i.to_h }
           ks_match = _array_unsorted_eql(remote_key_schema, lgsi[:key_schema])
 
+          # Throughput Check: Dependent on Billing Mode
           rpt = rgsi.provisioned_throughput.to_h
           lpt = lgsi[:provisioned_throughput]
-          pt_match = lpt.all? do |k,v|
-            rpt[k] == v
+          if @billing_mode == "PROVISIONED"
+            pt_match = lpt.all? do |k,v|
+              rpt[k] == v
+            end
+          elsif @billing_mode == "PAY_PER_REQUEST"
+            pt_match = lpt.nil? ? true : false
+          else
+            raise ArgumentError, "Unsupported billing mode #{@billing_mode}"
           end
 
           rp = rgsi.projection.to_h
@@ -537,10 +625,13 @@ module Aws
         if model_gsis
           model_gsis.each do |mgsi|
             config = gsi_config[mgsi[:index_name]]
-            # Validate throughput exists? Validate each throughput is in model?
-            gsis << mgsi.merge(
-              provisioned_throughput: config.provisioned_throughput
-            )
+            if @billing_mode == "PROVISIONED"
+              gsis << mgsi.merge(
+                provisioned_throughput: config.provisioned_throughput
+              )
+            else
+              gsis << mgsi
+            end
           end
         end
         gsis
@@ -553,8 +644,14 @@ module Aws
       def _validate_required_configuration
         missing_config = []
         missing_config << 'model_class' unless @model_class
-        missing_config << 'read_capacity_units' unless @read_capacity_units
-        missing_config << 'write_capacity_units' unless @write_capacity_units
+        if @billing_mode == "PROVISIONED"
+          missing_config << 'read_capacity_units' unless @read_capacity_units
+          missing_config << 'write_capacity_units' unless @write_capacity_units
+        else
+          if @read_capacity_units || @write_capacity_units
+            raise ArgumentError.new("Cannot have billing mode #{@billing_mode} with provisioned capacity.")
+          end
+        end
         unless missing_config.empty?
           msg = missing_config.join(', ')
           raise Errors::MissingRequiredConfiguration, 'Missing: ' + msg